+++ server/Allay/src/main/kotlin/cn/huohuas001/huhobot/allay/HuHoBotAllay.kt
@@ -154,6 +154,10 @@ class HuHoBotAllay: Plugin(), HuHoBot {
         return getPluginContainer().descriptor().version
     }
 
+    override fun getCallbackConvertImg(): Int {
+        return config.callbackConvertImg
+    }
+
     override fun loadCustomCommand() {
         val commands = config.customCommandMap